Calendar fin texts in styles
